Rewrite: Move to concept of genotype
For swine and human flu, the observed viruses have settled into known genotypes.
A genotype here is a named combination of segments, e.g. HA = pdm, NA = pdm, NS = avian etc.
The genetic makeup of a virus can then be succinctly expressed by its genotype: By saying something is a H1N2dk virus, the clade of each of its 8 segments is known.
Sometimes virus appears not in a known genotype due to ressortment, or if the sequence of any segment is truly unknown.
We also want warnings if that happens.

The following commit includes these changes and more:
* Re-organize the refdir/phylo/host:
  - segment dir contains e.g. HA/H1N1.fna and HA/H3N2.fna files
  - genotypes.tsv contains a list of known genotypes
  - tree_segments.csv contains a list of segments to make trees for.
* Genotypes and tree segments are loaded in beginning of phylo pipeline
* A genotypes.txt report is created after parsing BLAST