Change rules for matching samples with genotypes
Logic is now slightly more complicated. For each sample/genotype match, we
check if the pair is definitely a match (i.e. a Match for all considered
segments), could be a match (ambiguous BLAST hit, or missing segment, or any
of multiple copies match), or is definitely not a match (all copies have no
hits on BLAST or hit a different clade).

Then:
* A sample is empty if it has no segments
* If any segment copy has no BLAST hits, call it a new genotype
* If a genotype is a unique match, there can be no other compatible genotypes,
  error if not true.
* If unique match, report it as known genotype
* Else, if no compatible genotypes, it's a new genotype
* Else, report as ambiguous genotype