Create trees based on groups of clades
Lots of changes here.
Previously, assignment of clade was seen to sometimes have edge cases.
In order to give more information to make judgement calls on subtype
especially for swine viruses, it's best if viruses are placed in trees
with larger groups of clades, e.g. all H1 HA segments.

For this, we introduce the concept of a "group": A named set of segtypes
for a given segment. The name must be unique within a given segment.

The groups are read from a reference text file.

Guide trees are now created for all host species, for all segments that
belong to any listed groups.

This replaces the previous "tree segments" list.