Create trees for segments with ambiguous hits
Previously, only segcopies assigned clades were written to FASTA files to
be part of phylogenetic trees. However, the trees are most useful precisely
when dealing with ambiguous clades.

Now, an ambiguous hit will print its two closest clades in the report, and
be considered part of whatever tree groups the closest hit is.
It is not too important if the tree group is mis-assigned: The tree groups
should be large, and the tree should reveal their true identity.